# api/gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Tekton Hub
  description: HTTP services for managing Tekton Hub
  version: "1.0"
host: api.hub.tekton.dev
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /catalog/{catalogName}/error:
    get:
      tags:
      - catalog
      summary: CatalogError catalog
      description: List all errors occurred refreshing a catalog
      operationId: catalog#CatalogError
      parameters:
      - name: catalogName
        in: path
        description: Name of catalog
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/CatalogCatalogErrorResponseBody'
            required:
            - data
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/CatalogCatalogErrorInternalErrorResponseBody'
      schemes:
      - https
  /catalog/{catalogName}/refresh:
    post:
      tags:
      - catalog
      summary: Refresh catalog
      description: |-
        Refresh a Catalog by it's name

        **Required security scopes for jwt**:
          * `catalog:refresh`
      operationId: catalog#Refresh
      parameters:
      - name: catalogName
        in: path
        description: Name of catalog
        required: true
        type: string
      - name: Authorization
        in: header
        description: JWT
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/CatalogRefreshResponseBody'
        "404":
          description: Not Found response.
          schema:
            $ref: '#/definitions/CatalogRefreshNotFoundResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/CatalogRefreshInternalErrorResponseBody'
      schemes:
      - https
      security:
      - jwt_header_Authorization: []
  /catalog/refresh:
    post:
      tags:
      - catalog
      summary: RefreshAll catalog
      description: |-
        Refresh all catalogs

        **Required security scopes for jwt**:
          * `catalog:refresh`
      operationId: catalog#RefreshAll
      parameters:
      - name: Authorization
        in: header
        description: JWT
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: array
            items:
              $ref: '#/definitions/JobResponse'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/CatalogRefreshAllInternalErrorResponseBody'
      schemes:
      - https
      security:
      - jwt_header_Authorization: []
...
definitions:
...
  CatalogCatalogErrorInternalErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: Internal Server Error (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  CatalogCatalogErrorResponseBody:
    title: CatalogCatalogErrorResponseBody
    type: object
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/CatalogErrorsResponseBody'
        description: Catalog errors
        example:
        - errors:
          - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
          - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
          type: warning
        - errors:
          - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
          - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
          type: warning
        - errors:
          - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
          - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
          type: warning
    example:
      data:
      - errors:
        - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
        - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
        type: warning
      - errors:
        - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
        - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
        type: warning
      - errors:
        - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
        - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
        type: warning
    required:
    - data
  CatalogErrorsResponseBody:
    title: CatalogErrorsResponseBody
    type: object
    properties:
      errors:
        type: array
        items:
          type: string
          example: Omnis aut beatae reiciendis accusantium distinctio.
        description: Catalog Error message
        example:
        - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
        - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
      type:
        type: string
        description: Catalog Errror type
        example: warning
    description: CatalogErrors define the errors that occurred during catalog refresh
    example:
      errors:
      - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
      - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
      type: warning
    required:
    - type
    - errors
  CatalogRefreshAllInternalErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: Internal Server Error (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  CatalogRefreshInternalErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: Internal Server Error (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  CatalogRefreshNotFoundResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: Resource Not Found Error (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  CatalogRefreshResponseBody:
    title: 'Mediatype identifier: application/vnd.hub.job; view=default'
    type: object
    properties:
      catalogName:
        type: string
        description: Name of the catalog
        example: tekton
      id:
        type: integer
        description: id of the job
        example: 1
        format: int64
      status:
        type: string
        description: status of the job
        example: queued
    description: RefreshResponseBody result type (default view)
    example:
      catalogName: tekton
      id: 1
      status: queued
    required:
    - id
    - catalogName
    - status
...
  JobResponse:
    title: 'Mediatype identifier: application/vnd.hub.job; view=default'
    type: object
    properties:
      catalogName:
        type: string
        description: Name of the catalog
        example: tekton
      id:
        type: integer
        description: id of the job
        example: 1
        format: int64
      status:
        type: string
        description: status of the job
        example: queued
    description: The Job type describes a catalog refresh job that is run asynchronously
      (default view)
    example:
      catalogName: tekton
      id: 1
      status: queued
    required:
    - id
    - catalogName
    - status
...
securityDefinitions:
  jwt_header_Authorization:
    type: apiKey
    description: |-
      Secures endpoint by requiring a valid JWT retrieved via the /auth/login endpoint.

      **Security Scopes**:
        * `rating:read`: Read-only access to rating
        * `rating:write`: Read and write access to rating
        * `agent:create`: Access to create or update an agent
        * `catalog:refresh`: Access to refresh catalog
        * `config:refresh`: Access to refresh config file
        * `refresh:token`: Access to refresh user access token
    name: Authorization
    in: header